+++ src/components/dashboard/charts/CircleChart.tsx
@@ -1,9 +1,13 @@
 import React, { type FC } from "react"
-import { format } from "../data"
+import { formatLong } from "../data"
 
 interface Props {
-  value: number
-  comparative: number
+  data: {
+    value: number
+    color: string
+    description: string
+  }[]
+  displayValue: number
   unitOfMeasurement: string
 }
 
@@ -11,11 +15,12 @@ const CircleChart: FC<Props> = (props) => {
   const w = 60
   const h = 60
 
-  const t = 5
+  const r1 = 20
+  const r2 = 30
 
-  const total = props.value + props.comparative
+  const total = props.data.reduce((accumulator, v) => accumulator + v.value, 0)
 
-  if (props.value === 0 && props.comparative === 0) {
+  if (props.data.length === 0) {
     return (
       <svg viewBox={`0 0 ${w} ${h}`}>
         <rect x="0" y="0" width={w} height={h} fill="rgba(0,0,0,0.08)"></rect>
@@ -32,24 +37,26 @@ const CircleChart: FC<Props> = (props) => {
     )
   }
 
+  let baseline = r1
+
   return (
     <svg viewBox={`0 0 ${w} ${h}`}>
-      <circle
-        cx={w / 2}
-        cy={h / 2}
-        r={w / 3 + t / 2 - (t * props.value) / total / 2}
-        fill="none"
-        stroke="#fff"
-        strokeWidth={(t * props.value) / total}
-      />
-      <circle
-        cx={w / 2}
-        cy={h / 2}
-        r={w / 3 - t / 2 + (t * props.comparative) / total / 2}
-        fill="none"
-        stroke="hsl(0,0%,70%)"
-        strokeWidth={(5 * props.comparative) / total}
-      />
+      {props.data.map((d) => {
+        const currentBaseline = baseline
+        const strokeWidth = (d.value / total) * (r2 - r1)
+        const r = currentBaseline + strokeWidth / 2
+        baseline += strokeWidth
+        return (
+          <circle
+            cx={w / 2}
+            cy={h / 2}
+            r={r}
+            fill="none"
+            stroke={d.color}
+            strokeWidth={strokeWidth}
+          />
+        )
+      })}
       <text
         x={w / 2}
         y={h / 2}
@@ -57,11 +64,11 @@ const CircleChart: FC<Props> = (props) => {
         textAnchor="middle"
         style={{ fontSize: 8 }}
       >
-        {format(props.value)}
+        {formatLong(props.displayValue)}
       </text>
       <text
         x={w / 2}
-        y={h / 2 + 5}
+        y={h / 2 + 7}
         fill="#fff"
         textAnchor="middle"
         style={{ fontSize: 4 }}
